requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
drug2cell>=0.1.0
scanpy>=1.9.3
plotly>=5.13.0
imbalanced-learn>=0.10.1
statsmodels>=0.14.0
scipy>=1.11.0
scikit-learn>=1.3.0
tqdm>=4.64.0
xgboost>=1.7.0
PyYAML>=6.0
kaleido